refactor(import operations): structure: distinguish between import and comparison
